product implementation

+++ cashierApp/src/main/java/alimentation/cashierApp/services/ProductService.java
@@ -3,22 +3,33 @@ package alimentation.cashierApp.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import alimentation.cashierApp.dao.ProductRepository;
+import alimentation.cashierApp.exception.CashierAppException;
 import alimentation.cashierApp.models.Product;
+import alimentation.cashierApp.models.Transaction;
 
 @Service
-public class ProductService {
+public class ProductService { //requires error catching
 
     @Autowired
     private ProductRepository ProductRepository;
 
+    @Autowired
+    private TransactionService transactionService;
+
 
     // Add methods for Product-related operations here
-    Iterable<Product> getAllProducts(){
+    public Iterable<Product> getAllProducts(){
         return ProductRepository.findAll();
-    }    
+    }   
+
+    public Iterable<Product> getAllProductsByTransactionId(int transactionId){
+        Transaction transaction = transactionService.getTransactionById(transactionId);
+        return ProductRepository.findAllByTransaction(transaction);
+    }
     
     // filtering method is inefficient and poorly synergized with the transaction id, query directly from the dao
     // List<Product> getAllProductsByTransactionId(int transactionId){
@@ -32,7 +43,7 @@ public class ProductService {
     //     return products;
     // }
 
-    Product getProductById(int id){
+    public Product getProductById(int id){
         Optional<Product> response = ProductRepository.findById(id);
         if (response.isPresent()) {
             return response.get();
@@ -40,15 +51,21 @@ public class ProductService {
         return null;
     }
 
-    void addProduct(Product Product){
-        ProductRepository.save(Product);
+    public Product addProduct(Product Product){
+        return ProductRepository.save(Product);
     }
 
-    void updateProduct(Product Product){
-        ProductRepository.save(Product);
+    public Product updateProduct(Product Product){
+        return ProductRepository.save(Product);
     }
 
-    void deleteProduct(int id){
+    public Product deleteProduct(int id){
+        Optional<Product> target = ProductRepository.findById(id);
+        Product result = target.get();
+        if (result == null){
+            throw new CashierAppException(HttpStatus.BAD_REQUEST, "Product missing");
+        }
         ProductRepository.deleteById(id);
+        return result;
     }
 }
